+++ .ncurc.js
@@ -2,7 +2,5 @@
 
 module.exports = {
   reject: [
-    // Wait until ash-nazg can work with it and use flat config
-    'eslint-plugin-array-func'
   ]
 };